safety: add OnlySupported mode for convenience
Previously there were Mandatory and Try modes, but now we have
OnlySupported which only returns an error on locking if we are
running on a supported system (Linux). This should help cleanup
callers of landlock, so that we can use the interface directly
without using build tags (ignoring helpers for now).